day5/browserUser/requirements.txt
playwright
openai
python-dotenv 